requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0